templates: statement: added sim-statement.cls TeX class template
Statement template now consists of two files sim-statement.cls and statement.tex
where the former stores much of the boilerplate, so that statement.tex stays
concise.

With that support for referencing files relative to the TeX statement file
(doc/statement.tex) was added.

Moreover, ensuring that template does not replace an existing file is now done
for every template, not just the test generator.